+++ src/main/java/com/mx/mitienda/model/dto/DevolucionComprasRequestDTO.java
@@ -2,12 +2,14 @@ package com.mx.mitienda.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DevolucionComprasRequestDTO {
     private Long compraId;
     private Long branchId;
     private Long businessTypeId;
     private String codigoBarras;
-    private Integer cantidad;
+    private BigDecimal cantidad;
     private String motivo;
 }
