+++ src/main/java/com/mx/mitienda/model/dto/DevolucionVentasRequestDTO.java
@@ -2,11 +2,13 @@ package com.mx.mitienda.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DevolucionVentasRequestDTO {
     private Long ventaId;
     private String codigoBarras;
-    private Integer cantidad;
+    private BigDecimal cantidad;
     private String motivo;
     private Long branchId;
     private Long businessTypeId;
