+++ src/main/java/com/mx/mitienda/model/dto/HistorialMovimientosResponseDTO.java
@@ -3,14 +3,15 @@ package com.mx.mitienda.model.dto;
 import com.mx.mitienda.util.enums.TipoMovimiento;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class HistorialMovimientosResponseDTO {
     private LocalDateTime movementDate;
     private TipoMovimiento movementType;
-    private Integer quantity;
-    private Integer beforeStock;
-    private Integer newStock;
+    private BigDecimal quantity;
+    private BigDecimal beforeStock;
+    private BigDecimal newStock;
     private String reference;
 }
