+++ src/main/java/com/mx/mitienda/model/dto/InventarioAlertasDTO.java
@@ -2,13 +2,15 @@ package com.mx.mitienda.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InventarioAlertasDTO {
     private Long id;
     private String productName;
-    private Integer stock;
-    private Integer minStock;
-    private Integer maxStock;
+    private BigDecimal stock;
+    private BigDecimal minStock;
+    private BigDecimal maxStock;
     private Boolean stockCritico;
     private String branchName;
 }
