+++ src/main/java/com/mx/mitienda/model/dto/InventarioSucursalRequestDTO.java
@@ -3,13 +3,15 @@ package com.mx.mitienda.model.dto;
 import com.mx.mitienda.util.enums.InventarioOwnerType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InventarioSucursalRequestDTO {
     private Long productId;
     private Long branchId;
-    private Integer quantity;
-    private Integer minStock;
-    private Integer maxStock;
+    private BigDecimal quantity;
+    private BigDecimal minStock;
+    private BigDecimal maxStock;
     private Boolean isStockCritico;
     private InventarioOwnerType ownerType;
 }
