+++ src/main/java/com/mx/mitienda/model/dto/MovimientoStockDTO.java
@@ -3,10 +3,12 @@ package com.mx.mitienda.model.dto;
 import com.mx.mitienda.util.enums.InventarioOwnerType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MovimientoStockDTO {
     private Long productId;
-    private Integer cantidad;
+    private BigDecimal cantidad;
     private InventarioOwnerType ownerType;
 
 }
